pyproject.toml
[project]
name = "amc-backend"
version = "0.1.0"
description = "ASEAN Motor Club backend"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "django>=5.2.3",
    "django-ninja>=1.4.3",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.9",
    "uvicorn[standard]>=0.34.3",
    "uvicorn-worker>=0.3.0",
    "django-oauth-toolkit>=3.0.1",
    "aiohttp>=3.12.13",
    "django-asgi-lifespan>=0.4.0",
    "arq>=0.26.3",
    "factory-boy>=3.3.3",
    "discord-py>=2.5.2",
    "yarl>=1.20.1",
    "deepdiff>=8.5.0",
    "django-ninja-extra>=0.30.1",
    "cryptography>=45.0.5",
    "asyncstdlib>=3.13.1",
    "matplotlib>=3.10.6",
    "pycryptodome>=3.23.0",
    "psutil>=6.1.1",
    "pyjwt>=2.9.0",
]
[project.scripts]
amc-manage = "manage:main"

[build-system]
requires = ["hatchling", "editables"]
build-backend = "hatchling.build"

[tool.taskipy.tasks]
dev = "uvicorn amc_backend.asgi:application --reload --lifespan=on"
test = "django-admin test"

[tool.basedpyright]
allowedUntypedLibraries = [
  "django"
]

[tool.hatch.build.targets.wheel]
include = ["src/manage.py"]

[tool.hatch.build.targets.wheel.force-include]
"src" = "/"

[tool.pyrefly]
project_includes = [
    "**/*.py*",
    "**/*.ipynb",
]

[tool.uv.workspace]
members = [
    "packages/scripts",
]

[dependency-groups]
dev = [
    "basedpyright>=1.29.5",
    "django-stubs>=5.2.8",
    "pyrefly>=0.46.1",
    "ruff>=0.14.10",
    "taskipy>=1.14.1",
    "dpytest>=0.7.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-django>=4.9.0",
    "django-stubs-ext>=5.2.8",
]

